+++ src/Geoscape/BaseNameState.cpp
@@ -80,16 +80,16 @@ BaseNameState::BaseNameState(Base *base, Globe *globe, bool first, bool fixedLoc
 	if (!_game->getMod()->getBaseNamesFirst().empty())
 	{
 		std::ostringstream ss;
-		int pickFirst = RNG::seedless(0, _game->getMod()->getBaseNamesFirst().size() - 1);
+		int pickFirst = RNG::seedless(0, (int)_game->getMod()->getBaseNamesFirst().size() - 1);
 		ss << _game->getMod()->getBaseNamesFirst().at(pickFirst);
 		if (!_game->getMod()->getBaseNamesMiddle().empty())
 		{
-			int pickMiddle = RNG::seedless(0, _game->getMod()->getBaseNamesMiddle().size() - 1);
+			int pickMiddle = RNG::seedless(0, (int)_game->getMod()->getBaseNamesMiddle().size() - 1);
 			ss << " " << _game->getMod()->getBaseNamesMiddle().at(pickMiddle);
 		}
 		if (!_game->getMod()->getBaseNamesLast().empty())
 		{
-			int pickLast = RNG::seedless(0, _game->getMod()->getBaseNamesLast().size() - 1);
+			int pickLast = RNG::seedless(0, (int)_game->getMod()->getBaseNamesLast().size() - 1);
 			ss << " " << _game->getMod()->getBaseNamesLast().at(pickLast);
 		}
 		_edtName->setText(ss.str());
